improve graphics v4

See merge request morvaknin/trivia_gali_yuval_2024!67

+++ Server/Requests.h
@@ -60,7 +60,7 @@ public:
 	std::string roomName;
 	unsigned int maxUsers;
 	unsigned int questionCount;
-	double anwerTimeout;
+	unsigned int anwerTimeout;
 };
 
 struct SubmitAnswerRequest
